fixed warning: conversion of double to int.
git-svn-id: https://svn.lfbs.rwth-aachen.de/svn/mp-mpich/branches/rc-1.5@4994 3ddfa0b1-c7ec-0310-823c-dd91aa7df4cc

+++ mpid/ch_smi/smiinit.c
@@ -2023,7 +2023,7 @@ void MPID_SMI_Read_settings( void ) {
        has been choosen */
     if (MPID_SMI_cfg.USE_WATCHDOG == -MPID_SMI_WATCHDOG_DEF) {
 		MPID_SMI_cfg.USE_WATCHDOG = MPID_SMI_WATCHDOG_DEF 
-			+ MPID_SMI_numids*MPID_SMI_WATCHDOG_INC_PER_PROC;
+			+ (int)(MPID_SMI_numids*MPID_SMI_WATCHDOG_INC_PER_PROC);
 		/* This call may fail if SMI has watchdog disabled. */
 		SMI_Watchdog( MPID_SMI_cfg.USE_WATCHDOG);
     }
